# Conflicts:
#   pandora_console/include/help/clippy/extension_cron_send_email.php

+++ pandora_console/godmode/modules/manage_network_components_form_common.php
@@ -337,13 +337,13 @@ $next_row++;
 	//function paint graph
 	function paint_graph_values(){
 		//Parse integrer
-		var min_w = parseInt($('#text-min_warning').val());
+		var min_w = parseFloat($('#text-min_warning').val());
 			if(min_w == '0.00'){ min_w = 0; }
-		var max_w = parseInt($('#text-max_warning').val());
+		var max_w = parseFloat($('#text-max_warning').val());
 			if(max_w == '0.00'){ max_w = 0; }
-		var min_c = parseInt($('#text-min_critical').val());
+		var min_c = parseFloat($('#text-min_critical').val());
 			if(min_c =='0.00'){ min_c = 0; }
-		var max_c = parseInt($('#text-max_critical').val());
+		var max_c = parseFloat($('#text-max_critical').val());
 			if(max_c =='0.00'){ max_c = 0; }
 		var inverse_w = $('input:checkbox[name=warning_inverse]:checked').val();
 			if(!inverse_w){ inverse_w = 0; }
@@ -388,10 +388,10 @@ $next_row++;
 		//if haven't errors
 		if (error_w == 0 && error_c == 0){
 			//parse element
-			min_w = parseInt(min_w);
-			min_c = parseInt(min_c);
-			max_w = parseInt(max_w);
-			max_c = parseInt(max_c);
+			min_w = parseFloat(min_w);
+			min_c = parseFloat(min_c);
+			max_w = parseFloat(max_w);
+			max_c = parseFloat(max_c);
 			
 			//inicialize var
 			var range_min = 0;
